requirements.txt
google-genai
fastapi
uvicorn
python-dotenv
requests
beautifulsoup4
faiss-cpu
numpy
redis
sqlalchemy
psycopg2-binary
pandas
alembic
python-multipart
slowapi
twilio
websockets
edge-tts
pydub
scipy
pyyaml
